requirements.txt
flask
gunicorn
SQLAlchemy==1.3.16
Flask-SQLAlchemy==2.4.1
psycopg2-binary
jsonify
cloudinary
python-dotenv